Enable all warnings when compiling the C-versin of the game
Enabling the warnings revealed some .. warnings .. that were fixed,
mostly around the area of unused function-arguments.

+++ util.c
@@ -35,7 +35,7 @@ char *object_from_input(char *input)
 
     if (obj != NULL)
     {
-        for (int i = 0; i < strlen(obj); i++)
+        for (size_t i = 0; i < strlen(obj); i++)
         {
             obj[i] = tolower(obj[i]);
         }
